% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{chkchk}
\alias{chkchk}
\title{Exported function}
\usage{
chkchk(first, second)
}
\arguments{
\item{first, second}{Arguments}
}
\description{
A function that uses \code{chk_true()} on its first argument and calls an internal
function with the second argument.
The internal function also calls \code{chk_true()} on its argument.
}
